fix: add min of 10 sats to withdrawal, state check on tax change
This introduces a minimum to the withdrawal to prevent a panic when trying to calculate 10% of less than 10 sats. It also updates the state transition if exit tax is enabled or disabled so that the previous exit tax will not change unless the active change has gone into effect. You can reset the active change as many times as you want.